requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-multipart==0.0.5
pillow==9.0.0
numpy==1.21.0
scikit-learn==0.24.2
python-jose==3.3.0
pydantic==1.8.2
starlette==0.14.2
typings==3.7.4.3